Load & Run image API calls to run an image on current node
Distinguishes if the peerID given is itself or a remote peer.
