+++ p2p/upload_image.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"regexp"
 	"strconv"
 	"strings"
 	"time"
@@ -29,13 +28,11 @@ import (
 	"github.com/crowdcompute/crowdengine/log"
 
 	"github.com/crowdcompute/crowdengine/common"
+	"github.com/crowdcompute/crowdengine/common/dockerutil"
 	"github.com/crowdcompute/crowdengine/database"
-	"github.com/crowdcompute/crowdengine/manager"
 	api "github.com/crowdcompute/crowdengine/p2p/protomsgs"
 	uuid "github.com/satori/go.uuid"
 
-	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/api/types/filters"
 	host "github.com/libp2p/go-libp2p-host"
 	inet "github.com/libp2p/go-libp2p-net"
 	peer "github.com/libp2p/go-libp2p-peer"
@@ -110,7 +107,7 @@ func (p *UploadImageProtocol) onUploadRequest(s inet.Stream) {
 	err := createFileFromStream(s, filePath, fileSize)
 	common.FatalIfErr(err, "Couldn't read from stream when uploading a file")
 
-	imageID, err := loadImageToDocker(filePath)
+	imageID, err := dockerutil.LoadImageToDocker(filePath)
 	if errRemove := common.RemoveFile(filePath); errRemove != nil {
 		p.ImageIDchan <- errRemove.Error()
 		return
@@ -173,53 +170,6 @@ func createFileFromStream(s inet.Stream, toFilePath string, fileSize int64) erro
 	}
 	log.Println("File received completely!")
 	return nil
-}
-
-// loadImageToDocker takes a path to an image file and loads it to the docker daemon
-func loadImageToDocker(filePath string) (string, error) {
-	log.Println("Loading this image: ", filePath)
-	loadImageResp, err := manager.GetInstance().LoadImage(filePath)
-	if err != nil {
-		return "", err
-	}
-	log.Println(loadImageResp)
-	if imgID, exists := getImageID(loadImageResp); exists {
-		// Docker image ID is 64 characters
-		return imgID[:64], err
-	}
-	// If no image ID exists, we extract the image ID
-	// from listing the specific image using its tag
-	imageTag := loadImageResp[2 : len(loadImageResp)-5]
-	res := getImageSummaryFromTag(imageTag)
-	imgID := strings.Replace(res.ID, "sha256:", "", -1)
-	log.Println("Loaded image. Image ID: ", imgID)
-	return imgID, err
-}
-
-// imageIDExists checks if a docker image ID exists in the loadImageResp.
-// Docker image is just after the 'sha256:' prefix
-func getImageID(loadImageResp string) (string, bool) {
-	r, _ := regexp.Compile("sha256:(.*)")
-	matches := r.FindAllStringSubmatch(loadImageResp, -1)
-	if len(matches) != 0 {
-		return matches[0][1], true
-	}
-	return "", false
-}
-
-// getImageSummaryFromTag returns ImageSummaries from images using a tag
-func getImageSummaryFromTag(tag string) types.ImageSummary {
-	log.Println(tag)
-	fargs := filters.NewArgs()
-	fargs.Add("reference", tag)
-	res, err := manager.GetInstance().ListImages(
-		types.ImageListOptions{
-			Filters: fargs,
-		})
-	if err != nil {
-		log.Println("error: ", err)
-	}
-	return res[0] // we know that docker tag is unique thus returning only one summary
 }
 
 // storeImageToDB stores the new image's data to our level DB
